Changed address storage variable type from 'unsigned int' to 'unsigned long int' to avoid potential problem on 64-bit systems.

+++ idas/include/idaa.h
@@ -1,7 +1,7 @@
 /*
  * -----------------------------------------------------------------
- * $Revision: 1.17 $
- * $Date: 2006-03-24 15:57:23 $
+ * $Revision: 1.18 $
+ * $Date: 2006-06-12 18:45:37 $
  * ----------------------------------------------------------------- 
  * Programmers: Radu Serban @ LLNL
  * -----------------------------------------------------------------
@@ -161,11 +161,11 @@ extern "C" {
    * DEVELOPMENT USER-CALLABLE FUNCTIONS
    * ===============================================================
    */
-  
+
   /*
   typedef struct {
-    unsigned int my_addr;
-    unsigned int next_addr;
+    unsigned long int my_addr;
+    unsigned long int next_addr;
     realtype t0;
     realtype t1;
     long int nstep;
@@ -174,7 +174,7 @@ extern "C" {
   } IDAadjCheckPointRec;
 
   int IDAadjGetCheckPointsInfo(void *idaadj_mem, IDAadjCheckPointRec *ckpnt);
-  int IDAadjGetCurrentCheckPoint(void *idaadj_mem, unsigned int *addr);
+  int IDAadjGetCurrentCheckPoint(void *idaadj_mem, unsigned long int *addr);
 
   int IDAadjGetDataPointHermite(void *idaadj_mem, long int which,
                                 realtype *t, N_Vector y, N_Vector yd);
